requirements.txt
numpy>=1.26.0
sounddevice>=0.4.6
webrtcvad-wheels>=2.0.14
scipy>=1.11.0
faster-whisper>=1.0.3
ctranslate2>=4.3.1
spacy>=3.7.0
indic-nlp-library>=0.92
opencv-python>=4.9.0.80
mediapipe>=0.10.9
tqdm>=4.66.0
scikit-learn>=1.4.0
orjson>=3.10.0
joblib>=1.3.2
